+++ pio_usb_configuration.h
@@ -42,3 +42,6 @@ typedef struct {
 #define PIO_USB_ROOT_PORT_CNT 2
 
 #define PIO_USB_EP_SIZE 64
+
+// un-comment to use with TinyUSB
+// #define PIO_USB_USE_TINYUSB
